+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 ### Features
 
 * bem.scoped
+
+## v2.0.0
+
+### Features
+
+* The package is now an ESM module.
+* `bem.unscoped`, `bem.scoped`, and `bem.single` was deprecated.
+* The default export is now the `bemHelper` factory that returns a bem function with prebound parameters.
+* `bemHelper` supports CSS module styles via a second parameter (`styles`).